add: validator method, test version. add: jsonschema~=4.17.3 to requirements.txt add: test_validator.py
