+++ api_testing/utils/validator.py
@@ -1,13 +1,7 @@
 import jsonschema
-import requests
 import json
 from os.path import join, dirname
 from jsonschema import Draft202012Validator
-
-# End points
-BASE_PAGE = 'https://reqres.in'
-SINGLE_USER = '/api/users/2'
-LIST_USERS = '/api/users?page=2'
 
 
 def validator(data, schema_file):
@@ -37,19 +31,3 @@ def _load_json_schema(filename):
     except FileNotFoundError:
         print(f"No such JSON schema file: {relative_path}")
         return False
-
-
-# Tests
-class Tests:
-
-    def test_single_user(self):
-        response = requests.get(f'{BASE_PAGE}{SINGLE_USER}')
-        json_data = response.json()
-        print(json_data)
-        assert validator(json_data, 'single_user.json'), 'API response is incorrect'
-
-    def test_list_users(self):
-        response = requests.get(f'{BASE_PAGE}{LIST_USERS}')
-        json_data = response.json()
-        print(json_data)
-        assert validator(json_data, 'list_users.json'), 'API response is incorrect'
